+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 Changelog
 =========
 
+## TBD
+
+### Enhancements
+
+* The exception's `__notes__` field will now be sent as metadata if it exists
+  [#340](https://github.com/bugsnag/bugsnag-python/pull/340)
+
 ## v4.4.0 (2023-02-21)
 
 ### Enhancements