+++ CHANGELOG.md
@@ -8,6 +8,12 @@ Changelog
 * Allows changing the grouping hash when using the logger methods and
   [#334](https://github.com/bugsnag/bugsnag-python/pull/334)
 
+### Bug fixes
+
+* Fix the ignore class list not accounting for nested classes
+  [#342](https://github.com/bugsnag/bugsnag-python/pull/342)
+
+
 ## v4.4.0 (2023-02-21)
 
 ### Enhancements