+++ CHANGELOG.md
@@ -8,6 +8,9 @@ Changelog
 * Fixes one of the fields being mistakenly replaced with `[RECURSIVE]` when encoding a list or dictionary with identical siblings but no recursion.
   [#341](https://github.com/bugsnag/bugsnag-python/pull/341)
 
+* Fix the ignore class list not accounting for nested classes
+  [#342](https://github.com/bugsnag/bugsnag-python/pull/342)
+
 ## v4.4.0 (2023-02-21)
 
 ### Enhancements