+++ CHANGELOG.md
@@ -9,6 +9,10 @@ Changelog
   [#340](https://github.com/bugsnag/bugsnag-python/pull/340)
 
 ### Bug fixes
+
+* Fixes one of the fields being mistakenly replaced with `[RECURSIVE]` when encoding a list or dictionary with identical siblings but no recursion.
+  [#341](https://github.com/bugsnag/bugsnag-python/pull/341)
+
 * Fix the ignore class list not accounting for nested classes
   [#342](https://github.com/bugsnag/bugsnag-python/pull/342)