+++ CHANGELOG.md
@@ -8,6 +8,9 @@ Changelog
 * The exception's `__notes__` field will now be sent as metadata if it exists
   [#340](https://github.com/bugsnag/bugsnag-python/pull/340)
 
+* Allows changing the grouping hash when using `BugsnagHandler` via the logger methods' `extra` keyword argument
+  [#334](https://github.com/bugsnag/bugsnag-python/pull/334)
+
 ### Bug fixes
 
 * Fixes one of the fields being mistakenly replaced with `[RECURSIVE]` when encoding a list or dictionary with identical siblings but no recursion.